Tests for metrics on upload and downloading of files
